feat: most of job worker runner implementation plus a large blast radius
first off, `JobWorkerLauncher` was renamed to `JobWorkerRunner`. theres
a couple utility classes that got added `MachineIdentifier` and
`RunningCoordinator`. the `Process` model and table were also added.
this tracks, well, running processes. ok, ok, this is mostly for
heartbeat tracking to check if processes are alive or need to be
rotated. the supervisor's implementation still needs updated to check
heartbeat instead of relying on `Process.kill(0, pid)` which can be
unreliable for hung processes. the pipeline advancer also needs updated
to report a heartbeat