feat: re-implement `Ductwork::DefinitionBuilder`
im much happier with how this turned out vs the previous implementation.
somehow storing and manipulating a big ol hash felt more natural and
clean than dealing with obejcts for some reason.
another notable decision here is that we're structuring the hash as an
Adjacency List with edge metadata. the hope here is that this will make
pipeline advancement (DAG traversal) much easier even when there are
multiple parallel branches that can be running.
will take another commit or two to clean up before finishing
implementation and removing the `pending` on the specs