feat: implement expanding and collapsing in branch builder
i waited to refactor the `BranchBuilder` class beacuse it's starting to
smell a lot like the other builder class, `DefinitionBuilder`. it may
make sense to modify the initializer for `DefinitionBuilder` to take in
an existing definition and recursively build definitions. this should at
least mark being finished or nearly finished with the DSL implementation
so i can move on the pipeline advancement algorithm. the point of all
this change was to make traversal easier!