feat: implement `combine` action in pipeline advancer
this is the last of the pipeline advancement actions and represents a
big milestone: we can now run pipelines completely from end-to-end!!!
i've been testing this out with a toy app that has it installed via
local gem path. this project is desperate for test coverage around
anything that is run in a process. we'll get there