chore: basic refactor to make things a bit more readable
this also appeases rubocop a bit but now the class is too long lol.
admittedly, i am not happy with the current code quality. in the future,
i'll be establishing some patterns and do major refactors (hopefully).
one thing im specifically thinking about is not putting any methods on
the models and instead delegating to helper classes.