fix: make query for pipelines needing advancing uncached
for some reason AR was hitting its cache instead of executing the actual
query against the database. this was causing pipelines to only advance
when the process was started (since the cache is empty then). i havent
dug further into _why_ this was happening but i think this will suffice
for now (famous last words <sigh>). partially my thinking is this method
should be uncached anyway, maybe better to force it? iunno hopefully
theres time to come back to this