chore: move `logger` out of config to top-level `Ductwork` module
it makes more sense on the top-level `Ductwork` module as it's not
_really_ a configuration value. the logger level and source are actual
configuration values but the logger instance itself is a property of
`Ductwork`. it also makes it easier to call the logger because of less
typing lol.
technically, this is also a breaking change since this is sorta, kinda
part of the public interface but it's still early enough im not going to
cut a major release yet